make the controller anonymous
